+++ src/test/java/org/code_factory/jpa/nestedset/model/Category.java
@@ -5,15 +5,16 @@
 
 package org.code_factory.jpa.nestedset.model;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
 import org.code_factory.jpa.nestedset.NodeInfo;
 import org.code_factory.jpa.nestedset.annotations.LeftColumn;
 import org.code_factory.jpa.nestedset.annotations.LevelColumn;
 import org.code_factory.jpa.nestedset.annotations.RightColumn;
 import org.code_factory.jpa.nestedset.annotations.RootColumn;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
 
 /**
  * @author robo
@@ -21,7 +22,7 @@ import org.code_factory.jpa.nestedset.annotations.RootColumn;
 @Entity
 public class Category implements NodeInfo {
     @Id @GeneratedValue
-    private int id;
+    private Long id;
     private String name;
 
     @Column(updatable=false)
@@ -36,7 +37,7 @@ public class Category implements NodeInfo {
     @RootColumn
     private int rootId;
 
-    @Override public int getId() {
+    @Override public Long getId() {
         return this.id;
     }
